feat: improve max gain variable to store partial gain
ref #3

+++ src/structure/fix_result.rs
@@ -2,7 +2,7 @@
 pub(crate) struct UBFixResult {
     pub(crate) f_1: Vec<(usize, usize)>,
     pub(crate) x_best: Vec<[f64; 3]>,
-    pub(crate) v_best: i64,
+    pub(crate) v_best: f64,
 }
 
 impl UBFixResult {
@@ -10,7 +10,7 @@ impl UBFixResult {
         UBFixResult {
             f_1: vec![],
             x_best: vec![],
-            v_best: 0,
+            v_best: 0.0,
         }
     }
 }
