+++ src/structure/relaxation_result.rs
@@ -5,8 +5,8 @@ pub struct LPRelaxationResult {
     pub(crate) f_0: Vec<(usize, usize)>,
     pub(crate) x_up: Vec<[f64; 3]>,
     pub(crate) x: Vec<[f64; 3]>,
-    pub(crate) v_up: i64,
-    pub(crate) v_low: i64,
+    pub(crate) v_up: f64,
+    pub(crate) v_low: f64,
     pub(crate) relaxed: Vec<[ItemPreprocessing; 3]>,
 }
 
@@ -16,8 +16,8 @@ impl LPRelaxationResult {
             f_0: vec![],
             x_up: vec![],
             x: vec![],
-            v_up: 0,
-            v_low: 0,
+            v_up: 0.0,
+            v_low: 0.0,
             relaxed: vec![],
         }
     }
